update openharmony 1.0.1

+++ hiview_util.h
@@ -16,7 +16,7 @@
 #ifndef HOS_LITE_HIVIEW_UTIL_H
 #define HOS_LITE_HIVIEW_UTIL_H
 
-#include "hos_types.h"
+#include "ohos_types.h"
 
 #ifdef __cplusplus
 #if __cplusplus
@@ -52,7 +52,7 @@ typedef struct {
     uint16 millisecond;
 } HIVIEW_RtcTime;
 
-uint64 HIVIEW_GetCurrentTime(void);
+uint32 HIVIEW_GetCurrentTime(void);
 int32 HIVIEW_RtcGetCurrentTime(uint64 *val, HIVIEW_RtcTime *time);
 
 void *HIVIEW_MemAlloc(uint8 modId, uint32 size);
